+++ Sources/SwiftHTMLMacros/Plugin.swift
@@ -1,0 +1,9 @@
+import SwiftCompilerPlugin
+import SwiftSyntaxMacros
+
+@main
+struct SwiftHTMLMacrosPlugin: CompilerPlugin {
+    let providingMacros: [Macro.Type] = [
+        HTMLPreviewMacro.self,
+    ]
+}